fix: moving r1 to components
