+++ components/r1/index.js
